new philo_three : move checker inside each philo

+++ philo_three/srcs/parsing.c
@@ -6,7 +6,7 @@
 /*   By: mroux <mroux@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/14 18:53:15 by mroux             #+#    #+#             */
-/*   Updated: 2021/03/24 22:12:35 by mroux            ###   ########.fr       */
+/*   Updated: 2021/03/30 19:48:50 by mroux            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -50,7 +50,7 @@ int				extract_number_of_philosophers(int argc, char *argv[])
 	return (n);
 }
 
-unsigned long	extract_time_to_die(int argc, char *argv[])
+long long	extract_time_to_die(int argc, char *argv[])
 {
 	int		n;
 
@@ -60,7 +60,7 @@ unsigned long	extract_time_to_die(int argc, char *argv[])
 
 }
 
-unsigned long	extract_time_to_eat(int argc, char *argv[])
+long long	extract_time_to_eat(int argc, char *argv[])
 {
 	int		n;
 
@@ -70,7 +70,7 @@ unsigned long	extract_time_to_eat(int argc, char *argv[])
 
 }
 
-unsigned long	extract_time_to_sleep(int argc, char *argv[])
+long long	extract_time_to_sleep(int argc, char *argv[])
 {
 	int		n;
 
@@ -79,7 +79,7 @@ unsigned long	extract_time_to_sleep(int argc, char *argv[])
 	return (n);
 }
 
-unsigned long	extract_number_of_meals(int argc, char *argv[])
+long long	extract_number_of_meals(int argc, char *argv[])
 {
 	int		n;
